+++ templates/admin/website/delete_admin_website_template.sh
@@ -9,11 +9,14 @@ APACHE_SITES_ENABLED_DIR='SEDapache_sites_enabled_dirSED'
 APACHE_DOCROOT_DIR='SEDapache_docroot_dirSED'
 WEBSITE_VIRTUALHOST_CONFIG_FILE='SEDwebsite_virtualhost_fileSED' 
 WEBSITE_DOCROOT_ID='SEDwebsite_docroot_idSED'
-admin_log_file='/var/log/admin_install.log'
+
+script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+admin_log_file='/var/log/admin_website_install.log'
+
+cd "${script_dir}" || exit
 
 # Disable the admin site.
-cd /home/ec2-user || exit
-rm -f "${APACHE_SITES_ENABLED_DIR}"/"${WEBSITE_VIRTUALHOST_CONFIG_FILE}" 
+rm -f "${APACHE_SITES_ENABLED_DIR:?}"/"${WEBSITE_VIRTUALHOST_CONFIG_FILE:?}" 
 echo 'Admin Web Site virtualhost disabled' >> "${admin_log_file}"
 
 # Delete the Admin website
